requirements.txt
torch
torchvision
tensorflow
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
ultralytics
streamlit
Pillow
tqdm
